+++ src/main/scala/org/apache/spark/sql/redis/BinaryRedisPersistence.scala
@@ -16,7 +16,7 @@ class BinaryRedisPersistence extends RedisPersistence[Array[Byte]] {
   override def save(pipeline: Pipeline, key: String, value: Array[Byte], ttl: Int): Unit = {
     val keyBytes = key.getBytes(UTF_8)
     if (ttl > 0) {
-      pipeline.setex(keyBytes, ttl, value)
+      pipeline.setex(keyBytes, ttl.toLong, value)
     } else {
       pipeline.set(keyBytes, value)
     }
